Refactor Sonatype vulnerability API interface and handle auth
In this change, the `SonatypeVulnerabilityResponse.kt` file has been deleted and replaced with `SonatypeApis.kt`. This was done to facilitate not only vulnerability but also other types of responses that may come from the Sonatype APIs.

The security package has been renamed to sonatype-apis indicating that the package now handles all types of sonatype based api calls, not just security vulnerability related. The artifactId in the gradle build was also adjusted according to this change.

Furthermore, the new implementation now handles auth as evident from the `basic` method in the `defaultClient()` and the newly added `ktor-client-auth` dependency. This is an important addition for communicating with APIs that require authentication.

`SonatypeVulnerabilityRequest` was created to handle vulnerability requests in a more standardized and extensible way. Other different types of Sonatype API calls can similarly have their own request objects in the future.

Three different methods were added to the `SonatypeSecurityApiClient` to handle package vulnerabilities for different package managers (Maven, Cocoapods, Npm) using the same API call to Sonatype.

Overall, these changes provide a more extensible, clearer, and safer way of communicating with Sonatype's APIs.