Add workflow grouping to package publish action
A new group has been configured in the GitHub workflow to group all the jobs under the same GitHub head reference. Additionally, the cancel-in-progress flag has been enabled to ensure that if a newer workflow run is triggered, it will automatically cancel any previous runs still in progress. This change is implemented to reduce resource consumption and allow important workflows to run in a timely manner.