Add optional properties to Dependency class and normalize DefaultVersion class
Two properties (attributes and endorseStrictVersion) are added as optional to the Dependency class. Default values are added to the nullable properties of Semantic, TimestampLike, and Garbage sub-classes of the NormalizedVersion class. These changes aim to enrich the Dependency metadata and handle potential nullability and defaults in a superior way. Adjustments in PackageSearchApiClient functions were made accordingly.