Refactor and normalize codebase, adding default values to classes
This commit introduces a series of changes aimed at streamlining and increasing the readability of code. Import for `SearchParametersBuilder` and `buildSearchParameters` was added to `PackageSearchApiClient.kt` for improved search functionalities. The `defaultHttpClient` function was refactored for better configurability, allowing the necessary HTTP client configurations to be set in a more flexible way.

Furthermore, many classes like `ApiArtifact`, `ApiMavenPackage`, `VersionsContainer`, `GetScmByUrlRequest`, and a few others now have default values for their properties as opposed to previously being nullable. This will help to prevent null exceptions.

Finally, for the class `Vulnerability`, the `issues` property was switched from an empty list to nullable for better handling when no issues are present. These changes are geared towards modularization and improving code robustness and legibility.