Add code of conduct, contribution guide and modify README
The inclusion of the 'Contributor Covenant Code of Conduct' sets the behavioural guidelines for contribution to foster a healthy and respectful community. 'Contribution Guide' is added to streamline the contribution process and to inform contributors of expectations and processes. Also, README is being slightly modified to highlight the project's official JetBrains status.