Upgrade Gradle version and update various dependencies
Updated Gradle from 8.1.1 to 8.3 in 'gradle-wrapper.properties' to take advantage of new features and fixes in the recent build. Also, updated some packages in 'packagesearch-api-models.versions.toml' - 'kotlin' from 1.8.22 to 1.9.10, 'kotlinx-datetime' from 0.4.0 to 0.4.1, 'kotlinx-serialization' from 1.5.1 to 1.6.0 and 'ktor' from 2.3.3 to 2.3.4, for ensuring compatibility with the latest Kotlin version and to leverage improvements and bugfixes in these libraries.