Add RefreshPackagesInfo functionality to HTTP API
A new RefreshPackagesInfoRequest data class has been added to provide the necessary data for refreshing package information. This includes a list of packages with their ID and the latest known version. Furthermore, a new endpoint refreshPackagesInfo has been added to PackageSearchEndpoints and PackageSearchDefaultEndpoints. The PackageSearchApiClient now includes a suspend function refreshPackagesInfo to send the request.