Improve toString methods and refactor PackageSearchApiClient class
In this commit, toString methods were added to both ApiPackage and ApiRepository classes. Moreover, in the PackageSearchApiClient class, the obsolete constructor and several method signatures were refactored. After this, HttpRequestBuilder became optional for the majority of methods. The higher default value (from seconds to minutes) was also set to the isOnlineFlow function.